# Replace atomics with thread-local variables