--- a/Quake/gl_rmain.c
+++ b/Quake/gl_rmain.c
@@ -33,9 +33,7 @@ mplane_t frustum[4];
 qboolean render_warp;
 int      render_scale;
 
-// johnfitz -- rendering statistics
-atomic_uint32_t rs_brushpolys, rs_aliaspolys, rs_skypolys, rs_particles, rs_fogpolys;
-atomic_uint32_t rs_dynamiclightmaps, rs_brushpasses, rs_aliaspasses, rs_skypasses;
+THREAD_LOCAL per_thread_counters thread_counters;
 
 //
 // view origin
@@ -775,17 +773,31 @@ static void R_PrintStats (double time1)
 	// johnfitz -- modified r_speeds output
 	double time2 = 0;
 	if (r_speeds.value)
+	{
 		time2 = Sys_DoubleTime ();
+		for (int i = 0; i < Tasks_NumWorkers (); i++)
+		{ 
+			thread_counters.rs_brushpolys += thread_local_counters[i]->rs_brushpolys;
+			thread_counters.rs_brushpasses += thread_local_counters[i]->rs_brushpasses;
+			thread_counters.rs_aliaspolys += thread_local_counters[i]->rs_aliaspolys;
+			thread_counters.rs_aliaspasses += thread_local_counters[i]->rs_aliaspasses;
+			thread_counters.rs_dynamiclightmaps += thread_local_counters[i]->rs_dynamiclightmaps;
+			thread_counters.rs_skypolys += thread_local_counters[i]->rs_skypolys;
+			thread_counters.rs_skypasses += thread_local_counters[i]->rs_skypasses;
+		}
+	}
 	if (r_pos.value)
 		Con_Printf (
 			"x %i y %i z %i (pitch %i yaw %i roll %i)\n", (int)cl.entities[cl.viewentity].origin[0], (int)cl.entities[cl.viewentity].origin[1],
 			(int)cl.entities[cl.viewentity].origin[2], (int)cl.viewangles[PITCH], (int)cl.viewangles[YAW], (int)cl.viewangles[ROLL]);
 	else if (r_speeds.value == 2)
 		Con_Printf (
-			"%6.3f ms  %4u/%4u wpoly %4u/%4u epoly %3u lmap %4u/%4u sky\n", (time2 - time1) * 1000.0, rs_brushpolys, rs_brushpasses, rs_aliaspolys,
-			rs_aliaspasses, rs_dynamiclightmaps, rs_skypolys, rs_skypasses);
+			"%6.3f ms  %4u/%4u wpoly %4u/%4u epoly %3u lmap %4u/%4u sky\n", (time2 - time1) * 1000.0,
+			thread_counters.rs_brushpolys, thread_counters.rs_brushpasses, thread_counters.rs_aliaspolys, thread_counters.rs_aliaspasses,
+			thread_counters.rs_dynamiclightmaps, thread_counters.rs_skypolys, thread_counters.rs_skypasses);
 	else if (r_speeds.value)
-		Con_Printf ("%3i ms  %4i wpoly %4i epoly %3i lmap\n", (int)((time2 - time1) * 1000), rs_brushpolys, rs_aliaspolys, rs_dynamiclightmaps);
+		Con_Printf ("%3i ms  %4i wpoly %4i epoly %3i lmap\n", (int)((time2 - time1) * 1000),
+			thread_counters.rs_brushpolys, thread_counters.rs_aliaspolys, thread_counters.rs_dynamiclightmaps);
 	// johnfitz
 }
 
@@ -796,8 +808,7 @@ R_RenderView
 */
 void R_RenderView (qboolean use_tasks, task_handle_t begin_rendering_task, task_handle_t setup_frame_task, task_handle_t draw_done_task)
 {
-	static qboolean stats_ready;
-	double          time1;
+	double time1;
 
 	if (!cl.worldmodel)
 		Sys_Error ("R_RenderView: NULL worldmodel");
@@ -806,25 +817,12 @@ void R_RenderView (qboolean use_tasks, task_handle_t begin_rendering_task, task_
 	if (r_speeds.value)
 		time1 = Sys_DoubleTime ();
 
-	if (use_tasks && (r_pos.value || stats_ready))
+	if (use_tasks)
 		R_PrintStats (time1); // time2 will be ~= time1
 
-	if (r_speeds.value)
-	{
-		// johnfitz -- rendering statistics
-		Atomic_StoreUInt32 (&rs_brushpolys, 0u);
-		Atomic_StoreUInt32 (&rs_aliaspolys, 0u);
-		Atomic_StoreUInt32 (&rs_skypolys, 0u);
-		Atomic_StoreUInt32 (&rs_particles, 0u);
-		Atomic_StoreUInt32 (&rs_fogpolys, 0u);
-		Atomic_StoreUInt32 (&rs_dynamiclightmaps, 0u);
-		Atomic_StoreUInt32 (&rs_aliaspasses, 0u);
-		Atomic_StoreUInt32 (&rs_skypasses, 0u);
-		Atomic_StoreUInt32 (&rs_brushpasses, 0u);
-		stats_ready = true;
-	}
-	else
-		stats_ready = false;
+	memset (&thread_counters, 0, sizeof (per_thread_counters) - sizeof (uint32_t) * (MAX_SANITY_LIGHTMAPS - lightmap_count));
+	for (int i = 0; i < Tasks_NumWorkers (); i++) // this has to be done always, since it tracks lightmap use
+		memset (thread_local_counters[i], 0, sizeof (per_thread_counters) - sizeof (uint32_t) * (MAX_SANITY_LIGHTMAPS - lightmap_count));
 
 	cb_context_t *primary_cbx = &vulkan_globals.primary_cb_context;
 	if (use_tasks)