On a 6c/12t Skylake this appears to be quite a bit faster in some cases (1.23 -> 1.01ms in alk_dancing start, 12.12 -> 10.81ms in shib8 start, no change in ad_tears).