@Novum 2022-10-08
Fixed by 06e9fb789a2efd57ca196a293f84f18e7a0ef19b with simpler solution